+++ src/backend/resolvers/mutations/importWorkbook/utils/validate/tests/dupeRows/mockData/output/errors.js
@@ -1,6 +1,6 @@
 module.exports = [
   {
-    "error": "Row 4 is duplicated on row(s) 6, 10",
+    "error": '#Duplication Row Error\nRow 4 is duplicated on row(s) 6, 10',
     "rowNum": 4,
     "datum": {
       "fruit": "apple",