+++ src/backend/resolvers/mutations/importWorkbook/utils/validate/tests/location/mockData/output3/formattedErrors.js
@@ -1,5 +1,5 @@
-module.exports = `Mock Workbook/Mock Sheet failed validation for the following rows:
+module.exports = `#Mock Workbook/Mock Sheet failed validation for the following rows:
 
-Row 5, key 'location' Invalid location; try another location string or blank out the cell
+#Row 5, key 'location' Invalid location; try another location string or blank out the cell
 
 `