+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `Orion`: Built out OBM Service and OBM Service Category Placards, queries, mutations.
 
+### Fixed
+- `Payer Import Tool`: Fixed import error formatting by adding a # delimiter to each line of error description
+
 ## [2.7.0] - Started: 5-26-20
 
 ### Added